changed offerings calc in pagescontroller#home from to_i to to_f

+++ app/controllers/pages_controller.rb
@@ -8,8 +8,8 @@ class PagesController < ApplicationController
     @stripe_fees = 0
     @dailyOfferings = Offering.all.where(created_at: Time.zone.now.beginning_of_day..Time.zone.now.end_of_day)
     @dailyOfferings.each do |offering|
-      @offeringsToday += offering.amount.to_i/100
-      @stripe_fees += ((offering.amount.to_i * 0.029/100) + 0.30).round(2)
+      @offeringsToday += offering.amount.to_f/100
+      @stripe_fees += ((offering.amount.to_f * 0.029/100) + 0.30).round(2)
     end
 
   end
